+++ HISTORY.md
@@ -1,3 +1,11 @@
+# 0.4.7
+
+Allow conversion of any `FlexiChain{T}` to `MCMCChains.Chains`, as long as `Symbol(::T)` is defined.
+Previously this conversion only worked for `T <: VarName`.
+
+When indexing into a `FlexiChain{Symbol}` with a `Symbol`, FlexiChains now allows the symbol to resolve to an `Extra`, if such a key exists.
+Previously this would only look up parameters.
+
 # 0.4.6
 
 Adds a new function, `FlexiChains.to_nt_and_stats`, which samplers can overload in order to automatically bundle samples into `FlexiChain{Symbol}`.